Fix rtc compressor buffer overrun
Fix, say to me, how you are fixing this?

Thanks to Giperion

+++ src/xrCore/Compression/rt_compressor.cpp
@@ -16,14 +16,14 @@ u32 rtc_csize(u32 in)
 
 u32 rtc_compress(void* dst, u32 dst_len, const void* src, u32 src_len)
 {
-    u32 out_size = dst_len;
+    u64 out_size = dst_len;
     int r = lzo1x_1_compress((const lzo_byte*)src, (lzo_uint)src_len, (lzo_byte*)dst, (lzo_uintp)&out_size, rtc_wrkmem);
     VERIFY(r == LZO_E_OK);
     return out_size;
 }
 u32 rtc_decompress(void* dst, u32 dst_len, const void* src, u32 src_len)
 {
-    u32 out_size = dst_len;
+    u64 out_size = dst_len;
     int r = lzo1x_decompress((const lzo_byte*)src, (lzo_uint)src_len, (lzo_byte*)dst, (lzo_uintp)&out_size, rtc_wrkmem);
     VERIFY(r == LZO_E_OK);
     return out_size;
